jarvis/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
psycopg2-binary>=2.9.9
anthropic>=0.25.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic-settings>=2.2.0
rapidfuzz>=3.9.0
